Add hidden choice for the QA pad
